Linux 5.15

- drm_dp_helper.h: Get DP_DP13_MAX_LINK_RATE back which was removed by
  commit 7d56927efac7 ("drm/dp: add a number of DP 2.0 DPCD definitions")
- xlnx_dptx.c: Fix drm_dp_link_train_clock_recovery_delay() which was
  updated commit 9e9866664456 ("drm/dp: Pass drm_dp_aux to
  drm_dp_link_train_clock_recovery_delay()")
- cortex_arm64_edac.c: Replace deprecated CPU-hotplug functions as was
  done commit e104d530f373 ("hwmon: Replace deprecated CPU-hotplug
  functions.")
- spi-nor: Keep spi_nor_unlock_all/spi_nor_is_locked/spi_nor_lock functions
  because the whole spi framework needs to be updated
- xilinx_jesd204b.c - Add fallthrough shortcuts to avoid compilation
  warnings
- media: Follow commit 0d346d2a6f54 ("media: v4l2-subdev: add subdev-wide
  state struct")
- xlnx_drv.c: Update driver based on commit f5ca8eb6f9bd ("drm/cma-helper:
  Implement mmap as GEM CMA object functions"), commit fe319919f40d
  ("drm/xlnx: Don't set struct drm_device.irq_enabled") and commit
  c85dce1b2e7e ("drm/xlnx: Initialize DRM driver instance with CMA helper
  macro")
- zynqmp_disp/zynqmp_dp: Update driver based on commit 9e9866664456
  ("drm/dp: Pass drm_dp_aux to drm_dp_link_train_clock_recovery_delay()")
- ina260-adc: use devm_iio_kfifo_buffer_setup instead of
  devm_iio_kfifo_allocate as was done based on commit 17395ce29921 ("iio:
  makke use of devm_iio_kfifo_buffer_setup() helper") and id from struct
  based on commit 15ea2878bfb2 ("iio: core: move @id from struct iio_dev to
  struct iio_dev_opaque")
- zynqmp-sha/secure:
  "s/\b__flush_cache_user_range\b/caches_clean_inval_user_pou/g;" based on
  commit fade9c2c6ee2 ("arm64: Rename arm64-internal cache maintenance
  functions") and commit a24d22b225ce ("crypto: sha - split sha.h into
  sha1.h and sha2.h")
- vsc8531: Remove ack_interrupt based on commit 30446ae4675c ("net: phy:
  mscc: remove the use of .ack_interrupt()") for PHY_ID_VSC8531_02 phy
- spi-cadence-quadspi.c: Use Xilinx version - merge with upstream should
  happen
- uvc_video/vcu: use mainline version only
- Change value of V4L2_CTRL_TYPE_HDR from 0x0200 to 0x400 (0x200 is
  already taken)
- There can be doubled dt binding for some drivers which needs to be
  converted
- xilinx-m2m: Align driver based on commit c1cc23625062 ("media:
  v4l2-async: Discourage use of v4l2_async_notifier_add_subdev")
- irqdesc.c: Enable EXPORT_SYMBOL(irq_to_desc) for using with
  zynq_remoteproc module
- microblaze: Align Kbuild/atomic.h and mmu_context_mm.h based on commit
  fc63a6e08a8c ("locking/atomic: arm: move to ARCH_ATOMIC") and commit
  2fd171be13f0 ("m68k: use asm-generic/mmu_context.h for no-op implementations")
- microblaze: Change __IRQ_STAT handling based on commit 9f112156f8da
  ("parisc: Remove bogus __IRQ_STAT macro")
- zynq: Adopt upstream nand driver with also changing reference from nand0
  to nfc0 in DTS files
- zynq: Disable GCC plugins in defconfig
- axi_enet: Use xilinx version with changes described in commit
  83216e3988cd ("of: net: pass the dst buffer to of_get_mac_address()")
- Disable CPU idle driver which is causing issues on Versal for now
- Update macb driver with patches (done by Harini):
  net: macb: Update newly added usrio for ZynqMP
  net: macb: Update tsu_clk handling for Versal
- "spi: spi-zynqmp-gqspi: Fix suspend resume failure" by Amit
- dwc3: Align with Fixes: 124b11cc4f62 ("usb: dwc3: xilinx: Remove the
  extra freeing of clocks")
- "spi: spi-zynqmp-gqspi: Fix erase failure when tx buswidth is 4" by Amit
- "spi: spi-zynqmp-gqspi: Fix resume failure" by Amit
- "mtd: spi-nor: Update flash lock/unlock feature" by Amit
- Fix Octal SPI by Amit
  - Adapted the mainline DDR solution and removed the internal solution
  - Merged the upstreamed DMA solution
  - Disable the periodic tuning solution temporarily

+++ drivers/usb/gadget/legacy/inode.c
@@ -498,7 +498,8 @@ static void ep_aio_complete(struct usb_ep *ep, struct usb_request *req)
 		iocb->private = NULL;
 		/* aio_complete() reports bytes-transferred _and_ faults */
 
-		iocb->ki_complete(iocb, req->actual ? req->actual : req->status,
+		iocb->ki_complete(iocb,
+				req->actual ? req->actual : (long)req->status,
 				req->status);
 	} else {
 		/* ep_copy_to_user() won't report both; we hide some faults */
@@ -1213,8 +1214,8 @@ dev_release (struct inode *inode, struct file *fd)
 static __poll_t
 ep0_poll (struct file *fd, poll_table *wait)
 {
-       struct dev_data         *dev = fd->private_data;
-       __poll_t                mask = 0;
+	struct dev_data         *dev = fd->private_data;
+	__poll_t                mask = 0;
 
 	if (dev->state <= STATE_DEV_OPENED)
 		return DEFAULT_POLLMASK;
